worker/pyproject.toml
[tool.poetry]
name = "osint-worker"
version = "0.1.0"
description = "OSINT Platform Worker"
authors = ["OSINT Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
celery = "^5.3.0"
redis = "^5.0.0"
sqlalchemy = "^2.0.25"
psycopg = {extras = ["binary"], version = "^3.1.0"}
structlog = "^24.1.0"
dnspython = "^2.5.0"
httpx = "^0.27.0"
minio = "^7.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
